requirements.txt
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
cloudinary==1.44.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
fonttools==4.59.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
